+++ osdk/src/commands/clippy.rs
@@ -2,8 +2,8 @@
 
 use std::process;
 
-use super::utils::{cargo, COMMON_CARGO_ARGS};
-use crate::{commands::utils::create_target_json, error::Errno, error_msg};
+use super::util::{cargo, COMMON_CARGO_ARGS};
+use crate::{commands::util::create_target_json, error::Errno, error_msg};
 
 pub fn execute_clippy_command() {
     let target_json_path = create_target_json();
