+++ osdk/src/main.rs
@@ -13,7 +13,7 @@ mod cli;
 mod commands;
 mod config_manager;
 mod error;
-mod utils;
+mod util;
 
 fn main() {
     // init logger
