+++ osdk/src/util.rs
